# spec/openapi.yaml
openapi: 3.0.1
info:
  title: Prusa Link Web
  description:
    This is a printer connect webservice. It is compatible with another 3D
    printer services, so many applications which use this service
  contact:
    email: developer@prusa3d.cz
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 2.0.0
...
security:
  - ApiKeyAuth: [developer]
paths:
...
  /api/printer:
    get:
      summary: Retrive the current printer state
      parameters:
        - in: query
          name: exclude
          description: A list of attributes to not return in the response.
          schema:
            type: array
            items:
              type: string
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  temperature:
                    $ref: "#/components/schemas/TemperatureState"
                  sd:
                    $ref: "#/components/schemas/SDState"
                  state:
                    $ref: "#/components/schemas/PrinterState"
                  telemetry:
                    oneOf:
                      - $ref: "#/components/schemas/TelemetryMini"
                      - $ref: "#/components/schemas/TelemetrySL1"
        403:
          $ref: "#/components/responses/Unauthorized"
  /api/printer/sd:
    get:
      summary: Retrieves the current state of the printer’s SD card.
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SDState"
        403:
          $ref: "#/components/responses/Unauthorized"
  /api/printer/printhead:
    post:
      summary: Control printhead movement in XYZ axes
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                jog:
                  type: object
                  properties:
                    command:
                      type: string
                      example: "jog"
                    x:
                      type: number
                      example: 10
                    y:
                      type: number
                      example: -5
                    z:
                      type: number
                      example: 0.02
                home:
                  type: object
                  properties:
                    command:
                      type: string
                      example: "home"
                    axes:
                      type: array
                      items:
                        type: string
                        example: ["x", "y"]
                feedrate:
                  type: object
                  properties:
                    command:
                      type: string
                      example: "home"
                    factor:
                      type: number
                      example: 105
      responses:
        204:
          description: No error
        400:
          $ref: "#/components/responses/BadRequest"
        409:
          $ref: "#/components/responses/Conflict"
  /api/printer/tool:
    post:
      summary: Control extruder
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                target:
                  type: object
                  properties:
                    command:
                      type: string
                      example: "target"
                    targets:
                      type: object
                      properties:
                        tool0:
                          type: number
                          example: 220
                offset:
                  type: object
                  properties:
                    command:
                      type: string
                      example: "offset"
                    offsets:
                      type: object
                      properties:
                        tool0:
                          type: number
                          example: 10
                select:
                  type: object
                  properties:
                    command:
                      type: string
                      example: "select"
                    tool:
                      type: string
                      example: "tool0"
                extrude:
                  type: object
                  properties:
                    command:
                      type: string
                      example: "extrude"
                    amount:
                      type: number
                      example: 5
                retract:
                  type: object
                  properties:
                    command:
                      type: string
                      example: "extrude"
                    amount:
                      type: number
                      example: -3
                flowrate:
                  type: object
                  properties:
                    command:
                      type: string
                      example: "flowrate"
                    amount:
                      type: number
                      example: 95
      responses:
        204:
          description: No error
        400:
          $ref: "#/components/responses/BadRequest"
        409:
          $ref: "#/components/responses/Conflict"
  /api/printer/bed:
    post:
      summary: Control bed temperature
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                command:
                  type: string
                  example: "target"
                target:
                  type: number
                  example: 75

      responses:
        204:
          description: No error
        400:
          $ref: "#/components/responses/BadRequest"
        409:
          $ref: "#/components/responses/Conflict"

  # MINI
  # - will probably not support this endpoint.
  # - On error sends plaintext error description in response body
  # - On error possibly sends Content-Location header with URL to help.prusa3d.com.
  # - PCL page /error will be ommited

  # SL1
  # - supports this endpoint
  # - Response is dependent on Accept header in request
  # - If no Accept header is present:
  #   - sends plaintext error description in response body (compatibility with older slicer)
  #   - sends Content-Location header with URL to <PCL IP>/error (in the future it may send URL to hlep.prusa3d.com)
  # - if Accept header is present (currently supports only Accept: application/json):
  #   - sends JSON with same content as this /api/printer/error endpoint (code, title, text, url)
  #    sends Content-Location header with URL to <PCL IP>/error (in the future it may send URL to hlep.prusa3d.com)
  /api/printer/error:
    get:
      summary: Retrive the error number and text. Not compatible with OctoPrint.
      responses:
        200:
          description: Custom endpoint to get error details. Client can see if printer is in error state by polling /api/printer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        403:
          $ref: "#/components/responses/Unauthorized"
...
components:
  schemas:
...
    TemperatureData:
      type: object
      properties:
        actual:
          type: number
          description: Current temperature
        target:
          type: number
          description: Target temperature
          nullable: true
        offset:
          type: number
          nullable: true
          description: Currenntly configured temperature offset to apply.
      example:
        actual: 22.5
        target: 50.0
        offset: 1.0
    TemperatureState:
      type: object
      properties:
        tool0:
          $ref: "#/components/schemas/TemperatureData"
        bed:
          $ref: "#/components/schemas/TemperatureData"
        chamber:
          $ref: "#/components/schemas/TemperatureData"
    SDState:
      type: object
      properties:
        ready:
          type: boolean
          example: false
    PrinterState:
      type: object
      properties:
        text:
          $ref: "#/components/schemas/State"
        flags:
          type: object
          properties:
            operational:
              type: boolean
              example: false
            paused:
              type: boolean
              example: false
            printing:
              type: boolean
              example: true
            cancelling:
              type: boolean
              example: false
            pausing:
              type: boolean
              example: false
            sdReady:
              type: boolean
              example: false
            error: # this will indicate printer has an error to show in /api/priner/error
              type: boolean
              example: false
            ready:
              type: boolean
              example: true
            closedOrError:
              type: boolean
              example: false
            busy:
              type: boolean
              example: false
...
    State:
      type: string
      enum: [
        "Operational",
        "Printing",
        "Pausing",
        "Paused",
        "Cancelling",
        "Error",
        "Offline",
        "Busy",
      ]
      example: "Operational"
    TelemetryMini:
      type: object
      properties:
        temp_bed:
          type: number
        temp_nozzle:
          type: number
        material:
          type: string
    TelemetrySL1:
      type: object
      properties:
        fanUvLed:
          type: integer
          example: 2000
        fanBlower:
          type: integer
          example: 3333
        fanRear:
          type: integer
          example: 1000
        coverClosed:
          type: boolean
          example: true
        tempAmbient:
          type: number
          example: 23.4
        tempCpu:
          type: number
          example: 46.9
        tempUvLed:
          type: number
          example: 52.6

    #JobFileFDM:
...
    Error:
      type: object
      properties:
        code:
          type: string
          description: Prusa error code. Mus be string if we will have printer with code for example 04
          example: "10108"
        title:
          type: string
          description: Prusa error text string with prefiled variable macros.
          example: "RESIN TOO LOW"
        text:
          type: string
          description: Prusa error text string with prefiled variable macros.
          example: "Measured resin volume 22.4 ml is lower than required for this print. Refill the tank and restart the print."
        url:
          type: string
          description: Link to the Prusa help page
          example: "https://help.prusa3d.com/en/10108/LHE3Q0I1"
      required:
        - title
        - text
    ResponsePlaintext:
      type: string
      default: "Project with same name has been already uploaded."
      description: Prusa error message (raw_message)
  headers:
...
    Content-Location:
      required: true
      schema:
        type: string
        default: "http://<PCL IP>/error"
      description: URI path to nice error page. We can send URI to help.prusa3d.com in the future.
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  responses:
...
    BadRequest: # 400
      description: Bad Request
      headers:
        Content-Location:
          $ref: "#/components/headers/Content-Location"
      content:
        text/plain:
          schema:
            $ref: "#/components/schemas/ResponsePlaintext"
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthorized: # 401
      description: Unauthorized
      headers:
        Content-Location:
          $ref: "#/components/headers/Content-Location"
      content:
        text/plain:
          schema:
            $ref: "#/components/schemas/ResponsePlaintext"
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    Conflict: # 409
      description: Conflict
      headers:
        Content-Location:
          $ref: "#/components/headers/Content-Location"
      content:
        text/plain:
          schema:
            $ref: "#/components/schemas/ResponsePlaintext"
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...